refactor(context-engine): isolate maintenance write fence
